+++ spec/service_layer/contract_spec.rb
@@ -19,6 +19,31 @@ RSpec.describe ServiceLayer::Contract do
     end
   end
 
+  describe '.property!' do
+    subject { described_module.new(properties) }
+
+    let(:properties) { { email: 'adriensldy@gmail.com' } }
+
+    before { described_module.property! :email }
+
+    it 'creates properties' do
+      is_expected.to respond_to(:email, :email=)
+    end
+
+    it 'assigns properties' do
+      is_expected.to have_attributes(properties)
+    end
+
+    context 'when a property is missing' do
+      before { described_module.property! :first_name }
+
+      it 'raises ArgumentError with "The :first_name property is missing"' do
+        message = 'The :first_name property is missing'
+        expect(&method(:subject)).to raise_error(ArgumentError, message)
+      end
+    end
+  end
+
   describe '#render' do
     subject { contract.render }