+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add an alias `.call` for `Command.perform`.
 * Implement implicit `Command` class casting to `Proc`.
 * Setup the exceptions to rescue during the command invoke.
+* Allow the `Contract` to set some properties as required ones.
 
 ## [0.4.0] (2018-01-30)